Created a second translation unit, movegen.cpp, storing all the function definitioins previously in core.hpp
